fixed CMD_DB_LIST for levels != 0
dmconfig_list_cb() did not properly finalize AVP groups,
resulting in a DM2_REQUEST with level > 0 at the end which
meant that the entire request was not finalized properly.

As a result, the dmconfig client would not be able to parse
the RC AVP from the rpc_db_list() request, returning RC_ERR_MISC
instead.